test(transaction-detail): consolidate explorer spec into row-actions spec
transaction-detail-explorer.spec.tsx (from #3946) and this PR's spec
duplicated the mock harness and the hex-hash explorer case. Fold the
spark-UUID and hex explorer-routing cases into the row-actions spec
and drop the redundant file — one spec, one harness, covering every
icon on the screen.

Refs #3732
